Processing user_id on the server, API auth middleware

+++ server/api/syncTodos.ts
@@ -20,6 +20,7 @@ export default defineEventHandler(async (event) => {
 
   return new Promise(async (resolve, reject) => {
     try {
+      const { userId } = event.context.auth;
       const syncedIds: Number[] = [];
 
       const updatedRows = await Promise.all(todos.map(async (todo: Todo) => {
@@ -32,7 +33,7 @@ export default defineEventHandler(async (event) => {
 
         if (deleted && id) {
           // Delete the row from the database
-          await db.run('DELETE FROM todos WHERE id = ?;', id);
+          await db.run('DELETE FROM todos WHERE id = ? AND user_id = ?;', id, userId);
 
           syncedIds.push(id);
 
@@ -43,7 +44,7 @@ export default defineEventHandler(async (event) => {
           }
         } else if (id === null) {
           // Create a new row in the database
-          const result = await db.query!('INSERT INTO todos (content, done) VALUES (?, ?) RETURNING id', [content, done]);
+          const result = await db.query!('INSERT INTO todos (content, done, user_id) VALUES (?, ?, ?) RETURNING id', [content, done, userId]);
 
           syncedIds.push(result.rows[0].id);
 
@@ -54,7 +55,7 @@ export default defineEventHandler(async (event) => {
           }
         } else {
           // Update the row in the database
-          await db.run('UPDATE todos SET content = ?, done = ? WHERE id = ?;', content, done, id);
+          await db.run('UPDATE todos SET content = ?, done = ? WHERE id = ? AND user_id = ?;', content, done, id, userId);
 
           syncedIds.push(id);
         }
@@ -68,7 +69,7 @@ export default defineEventHandler(async (event) => {
       const syncedIdsStringified = syncedIds.join(',');
 
       let otherRows: Todo[] = await new Promise((resolve, reject) => {
-        db.all(`SELECT * FROM todos WHERE ID NOT IN (${syncedIdsStringified})`, (error, rows: Todo[]) => {
+        db.all(`SELECT * FROM todos WHERE ID NOT IN (${syncedIdsStringified}) AND user_id = ?`, [userId], (error, rows: Todo[]) => {
           if (error) {
             reject(error);
           } else {
